+++ apps/canonical-bridge-server/src/common/constants/index.ts
@@ -27,17 +27,22 @@ export enum Queues {
 }
 
 export enum Tasks {
-  fetchToken = 'fetchToken',
-  fetchCoingeckoToken = 'fetchCoingeckoToken',
-  fetchPrice = 'fetchPrice',
+  fetchCmcTokens = 'fetchTokens',
+  fetchCoingeckoTokens = 'fetchCoingeckoTokens',
+
+  fetchCmcPrice = 'fetchCmcPrice',
   fetchLlamaPrice = 'fetchLlamaPrice',
+
+  cacheCmcConfig = 'cacheCmcConfig',
+  cacheLlamaConfig = 'cacheLlamaConfig',
+  cacheCmcConfigV2 = 'cacheCmcConfigV2',
+  cacheLlamaConfigV2 = 'cacheLlamaConfigV2',
+
   fetchCbridge = 'fetchCbridge',
   fetchDebridge = 'fetchDebridge',
   fetchStargate = 'fetchStargate',
   fetchMeson = 'fetchMeson',
   fetchLayerZero = 'fetchLayerZero',
-  cacheCmcConfig = 'cacheCmcConfig',
-  cacheLlamaConfig = 'cacheLlamaConfig',
 
   filterCBridge = 'filterCBridge',
   filterDeBridge = 'filterDeBridge',
@@ -50,8 +55,9 @@ export enum Tasks {
   fetchCmsGlobalConfig = 'fetchCmsGlobalConfig',
 }
 
-export const TOKEN_REQUEST_LIMIT = 1000;
-export const PRICE_REQUEST_LIMIT = 200;
+export const CMC_TOKEN_REQUEST_LIMIT = 2000;
+export const CMC_PRICE_REQUEST_LIMIT = 1000;
+export const LLAMA_PRICE_REQUEST_LIMIT = 100;
 
 export const CACHE_KEY = {
   TOKEN_STARTED: 'token:started',
@@ -68,6 +74,9 @@ export const CACHE_KEY = {
   CMC_CONFIG: 'cmc:config',
   LLAMA_CONFIG: 'llama:config',
   PLATFORM_MAPPING: 'llama:platform',
+
+  CMC_CONFIG_V2: 'cmc:config_v2',
+  LLAMA_CONFIG_V2: 'llama:config_v2',
 
   FIELDED_CBRIDGE_CONFIG: 'bridge:filtered:cbridge',
   FIELDED_DEBRIDGE_CONFIG: 'bridge:filtered:debridge',