+++ packages/canonical-bridge-widget/src/modules/aggregator/hooks/useSelection.ts
@@ -20,6 +20,8 @@ import { useBridgeConfig } from '@/index';
 import {
   setFromChain,
   setSelectedToken,
+  setSendValue,
+  setToAccount,
   setToChain,
   setToToken,
   setToTokens,
@@ -73,8 +75,10 @@ export function useSelection() {
         // eslint-disable-next-line no-console
         console.log('[aggregator]', `has multiple toTokens (${toTokens.length})`);
         dispatch(setToTokens(toTokens));
+        dispatch(setToToken(undefined));
       } else {
         dispatch(setToTokens([]));
+        dispatch(setToToken(undefined));
       }
 
       if (toTokens.length === 1) {
@@ -216,6 +220,9 @@ export function useSelection() {
   const exchange = async () => {
     if (!aggregator) return;
 
+    dispatch(setSendValue(''));
+    dispatch(setToAccount({ address: '' }));
+
     const fromChainId = toChain!.id;
     const toChainId = fromChain!.id;
 
@@ -317,7 +324,12 @@ function useSortedTokens() {
 
       const tmpTokens = tokens.map((item) => {
         const balance = balances[item.address?.toLowerCase()];
-        const price = getTokenPrice(item);
+        const price = getTokenPrice({
+          chainId: fromChainId,
+          chainType,
+          tokenAddress: item?.address,
+          tokenSymbol: item?.symbol,
+        });
 
         let value: number | undefined;
         if (balance !== undefined && price !== undefined) {