+++ packages/canonical-bridge-widget/src/modules/transfer/components/SendInput/MaxLink.tsx
@@ -20,7 +20,12 @@ export const MaxLink: React.FC = () => {
   const { getTokenPrice } = useTokenPrice();
 
   const balance = getTokenBalance(selectedToken);
-  const tokenPrice = getTokenPrice(selectedToken);
+  const tokenPrice = getTokenPrice({
+    chainId: fromChain?.id,
+    chainType: fromChain?.chainType,
+    tokenAddress: selectedToken?.address,
+    tokenSymbol: selectedToken?.symbol,
+  });
 
   const setMaxAmount = () => {
     if (!!balance && selectedToken) {